closeBrowser property added with default value "true"

+++ src/main/java/utils/TestSetup.java
@@ -14,16 +14,18 @@ import static callsMethods.Methods.openCXphone;
  */
 public class TestSetup {
     public static void setup(WebDriver driver) throws InterruptedException, FindFailed, IOException {
-        Runtime.getRuntime().exec("taskkill /F /IM 3CXPhone.exe");
-        Thread.sleep(2000); //might fix phone not opened problem
-        System.out.println("3CXPhone killed from setup method.");
-        String hostName = InetAddress.getLocalHost().getHostName();
+        if (Boolean.getBoolean("closeBrowser")) {
+            Runtime.getRuntime().exec("taskkill /F /IM 3CXPhone.exe");
+            Thread.sleep(2000); //might fix phone not opened problem
+            System.out.println("3CXPhone killed from setup method.");
+            String hostName = InetAddress.getLocalHost().getHostName();
 
             Runtime.getRuntime().exec("taskkill /F /IM iexplore.exe");
             if (!isLocal()) {
                 Runtime.getRuntime().exec("taskkill /F /IM chrome.exe");
             }
-        openCXphone(60);
-        System.out.println("openCXphone method called from setup method.");
+            openCXphone(60);
+            System.out.println("openCXphone method called from setup method.");
+        }
     }
 }
